pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-regression-scanner"
version = "1.0.0"
description = "TUI tool to detect visual regressions on websites via screenshot comparison"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Michael Blaess"}]
dependencies = [
    "textual>=0.40",
    "rich>=13.0",
    "playwright>=1.40",
    "httpx>=0.25",
    "Pillow>=10.0",
]

[project.scripts]
visual-regression-scanner = "visual_regression_scanner.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
